Moving search resource to its own place.
