+++ api/src/main/java/marquez/api/SearchResource.java
@@ -12,11 +12,7 @@ import com.codahale.metrics.annotation.ExceptionMetered;
 import com.codahale.metrics.annotation.ResponseMetered;
 import com.codahale.metrics.annotation.Timed;
 import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import java.io.IOException;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
@@ -36,25 +32,18 @@ import marquez.api.models.SearchFilter;
 import marquez.api.models.SearchResult;
 import marquez.api.models.SearchSort;
 import marquez.db.SearchDao;
-import marquez.service.SearchService;
-import marquez.service.ServiceFactory;
-import org.opensearch.client.opensearch.core.SearchResponse;
-import org.opensearch.client.opensearch.core.search.Hit;
 
 @Slf4j
-@Path("/api/")
+@Path("api/v1/search")
 public class SearchResource {
   private static final String YYYY_MM_DD = "^\\d{4}-\\d{2}-\\d{2}$";
   private static final String DEFAULT_SORT = "name";
   private static final String DEFAULT_LIMIT = "10";
   private static final int MIN_LIMIT = 0;
 
-  private final SearchService searchService;
   private final SearchDao searchDao;
 
-  public SearchResource(
-      @NonNull final ServiceFactory serviceFactory, @NonNull final SearchDao searchDao) {
-    this.searchService = serviceFactory.getSearchService();
+  public SearchResource(@NonNull final SearchDao searchDao) {
     this.searchDao = searchDao;
   }
 
@@ -63,7 +52,6 @@ public class SearchResource {
   @ExceptionMetered
   @GET
   @Produces(APPLICATION_JSON)
-  @Path("v1/search")
   public Response search(
       @QueryParam("q") @NotBlank String query,
       @QueryParam("filter") @Nullable SearchFilter filter,
@@ -82,48 +70,6 @@ public class SearchResource {
             toLocateDateOrNull(before),
             toLocateDateOrNull(after));
     return Response.ok(new SearchResults(searchResults)).build();
-  }
-
-  @Timed
-  @ResponseMetered
-  @ExceptionMetered
-  @GET
-  @Produces(APPLICATION_JSON)
-  @Path("v2beta/search/jobs")
-  public Response searchJobs(@QueryParam("q") @NotBlank String query) throws IOException {
-    return formatOpenSearchResponse(this.searchService.searchJobs(query));
-  }
-
-  @Timed
-  @ResponseMetered
-  @ExceptionMetered
-  @GET
-  @Produces(APPLICATION_JSON)
-  @Path("v2beta/search/datasets")
-  public Response searchDatasets(@QueryParam("q") @NotBlank String query) throws IOException {
-    return formatOpenSearchResponse(this.searchService.searchDatasets(query));
-  }
-
-  private Response formatOpenSearchResponse(SearchResponse<ObjectNode> response) {
-    List<ObjectNode> hits =
-        response.hits().hits().stream().map(Hit::source).collect(Collectors.toList());
-    List<Map<String, List<String>>> highlights =
-        response.hits().hits().stream().map(Hit::highlight).collect(Collectors.toList());
-
-    return Response.ok(new OpenSearchResult(hits, highlights)).build();
-  }
-
-  @ToString
-  public static final class OpenSearchResult {
-    @Getter private final List<ObjectNode> hits;
-    @Getter private final List<Map<String, List<String>>> highlights;
-
-    @JsonCreator
-    public OpenSearchResult(
-        @NonNull List<ObjectNode> hits, @NonNull List<Map<String, List<String>>> highlights) {
-      this.hits = hits;
-      this.highlights = highlights;
-    }
   }
 
   /** Wrapper for {@link SearchResult}s which also contains a {@code total count}. */
